Reorganize test structure
